refactor: logic services to logic directory
Took 38 seconds
