+++ src/main/java/dev/realtards/kuenyawz/boostrappers/DatabaseBootstrapper.java
@@ -6,7 +6,7 @@ import dev.realtards.kuenyawz.entities.Account;
 import dev.realtards.kuenyawz.exceptions.AccountExistsException;
 import dev.realtards.kuenyawz.repositories.ProductRepository;
 import dev.realtards.kuenyawz.services.entity.AccountService;
-import dev.realtards.kuenyawz.services.ProductCsvService;
+import dev.realtards.kuenyawz.services.logic.ProductCsvService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.CommandLineRunner;
