+++ src/main/java/dev/realtards/kuenyawz/controllers/AuthController.java
@@ -3,7 +3,7 @@ package dev.realtards.kuenyawz.controllers;
 import dev.realtards.kuenyawz.dtos.account.AccountRegistrationDto;
 import dev.realtards.kuenyawz.dtos.account.AccountSecureDto;
 import dev.realtards.kuenyawz.dtos.auth.*;
-import dev.realtards.kuenyawz.services.AuthService;
+import dev.realtards.kuenyawz.services.logic.AuthService;
 import dev.realtards.kuenyawz.services.entity.OTPService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
