+++ src/main/java/dev/realtards/kuenyawz/services/CartItemService.java
@@ -1,0 +1,4 @@
+package dev.realtards.kuenyawz.services;
+
+public interface CartItemService {
+}
