+++ src/main/java/dev/realtards/kuenyawz/services/logic/RecommenderServiceImpl.java
@@ -1,4 +1,4 @@
-package dev.realtards.kuenyawz.services;
+package dev.realtards.kuenyawz.services.logic;
 
 import dev.realtards.kuenyawz.dtos.product.ProductDto;
 import dev.realtards.kuenyawz.entities.Product;
