+++ src/test/java/dev/realtards/kuenyawz/boostrappers/DatabaseBootstrapperTest.java
@@ -2,7 +2,7 @@ package dev.realtards.kuenyawz.boostrappers;
 
 import dev.realtards.kuenyawz.repositories.ProductRepository;
 import dev.realtards.kuenyawz.services.entity.AccountService;
-import dev.realtards.kuenyawz.services.ProductCsvService;
+import dev.realtards.kuenyawz.services.logic.ProductCsvService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
