+++ src/test/java/dev/realtards/kuenyawz/services/ProductServiceImplTest.java
@@ -9,6 +9,7 @@ import dev.realtards.kuenyawz.mapper.ProductMapper;
 import dev.realtards.kuenyawz.mapper.VariantMapper;
 import dev.realtards.kuenyawz.repositories.ProductRepository;
 import dev.realtards.kuenyawz.services.entity.ProductServiceImpl;
+import dev.realtards.kuenyawz.services.logic.ImageStorageService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
